requirements.txt
# Web Crawling
crawl4ai>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.18.0

# Multi-Agent AI
crewai>=0.28.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# LLM Providers
openai>=1.12.0
ollama>=0.1.0
google-generativeai>=0.3.0
langchain-google-genai>=0.0.6

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.6.0

# Scheduling
apscheduler>=3.10.4

# Web Framework (for API/Dashboard)
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
tqdm>=4.66.0
